shopease site

+++ ShopeEase--main/ShopeEase/ShopEase/src/main/java/com/example/ShopEase/dto/CartDTO.java
@@ -1,6 +1,8 @@
 package com.example.ShopEase.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -8,5 +10,5 @@ public class CartDTO {
     private Long cartId;
     private Long userId;
     private List<CartItemDTO> items;
-    private double totalPrice;
+    private BigDecimal totalPrice;
 }
