+++ ShopeEase--main/ShopeEase/ShopEase/src/main/java/com/example/ShopEase/dto/CartItemDTO.java
@@ -2,12 +2,14 @@ package com.example.ShopEase.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDTO {
     private Long userId;
     private Long productId;
     private String productName;
     private int quantity;
-    private double pricePerUnit;
-    private double totalItemPrice;
+    private BigDecimal pricePerUnit;
+    private BigDecimal totalItemPrice;
 }
