+++ ShopeEase--main/ShopeEase/ShopEase/src/main/java/com/example/ShopEase/dto/OrderItemDTO.java
@@ -2,10 +2,12 @@ package com.example.ShopEase.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderItemDTO {
     private String productName;
     private int quantity;
-    private double price;
-    private double totalPrice;
+    private BigDecimal price;
+    private  BigDecimal totalPrice;
 }
